1_data_ingestion_service/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
sqlalchemy
grpcio
protobuf
pytest
flake8
fastapi_pagination
pydantic
pydantic-core
pydantic-extra-types
pydantic-settings
asyncpg
aiohttp
graypy
greenlet
aiosqlite